Deangularization part 2
