+++ src/app/bungie-api/bungie-core-api.ts
@@ -1,5 +1,13 @@
 import { bungieApiQuery } from './bungie-api-utils';
 import { $http } from 'ngimport';
+import { IPromise } from 'angular';
+
+export interface GlobalAlert {
+  key: string;
+  type: string;
+  body: string;
+  timestamp: string;
+}
 
 // http://destinydevs.github.io/BungieNetPlatform/docs/Enums
 const GlobalAlertLevelsToToastLevels = [
@@ -12,9 +20,9 @@ const GlobalAlertLevelsToToastLevels = [
 /**
  * Get global alerts (like maintenance warnings) from Bungie.
  */
-export function getGlobalAlerts() {
+export function getGlobalAlerts(): IPromise<GlobalAlert[]> {
   return $http(bungieApiQuery(`/Platform/GlobalAlerts/`))
-    .then((response) => {
+    .then((response: any) => {
       if (response && response.data && response.data.Response) {
         return response.data.Response.map((alert) => {
           return {
