+++ src/app/dimApp.module.js
@@ -17,7 +17,6 @@ import 'angular-hotkeys';
 import 'angular-promise-tracker';
 import ocLazyLoadModule from 'oclazyload';
 
-import bungieApiModule from './bungie-api/bungie-api.module';
 import accountsModule from './accounts/accounts.module';
 import { ShellModule } from './shell/shell.module';
 import inventoryModule from './inventory/inventory.module';
@@ -55,7 +54,6 @@ const dependencies = [
   SliderModule,
   ToasterModule,
   UIRouterModule,
-  bungieApiModule,
   accountsModule,
   inventoryModule,
   itemReviewModule,
