+++ src/app/record-books/record-books.module.js
@@ -1,11 +1,10 @@
 import angular from 'angular';
 import 'angular-duration-format';
 
-import bungieApiModule from '../bungie-api/bungie-api.module';
 import { RecordBooksComponent } from './record-books.component';
 
 export default angular
-  .module('recordBooksModule', ['angular-duration-format', bungieApiModule])
+  .module('recordBooksModule', ['angular-duration-format'])
   .component('recordBooks', RecordBooksComponent)
   .config(($stateProvider) => {
     'ngInject';
